+++ src/app/api/resume/[candidateId]/experience/[roleId]/[accomplishmentId]/route.ts
@@ -2,7 +2,7 @@
 
 import { type NextRequest, NextResponse } from "next/server"
 import { validateDataFound, validateUuidFormat } from "@/lib/api/validate"
-import { getCandidateByCandidateId } from "@/lib/db/resume/candidate"
+import { getCandidateByCandidateId } from "@/lib/db/resume/candidate/candidate"
 import { getAccomplishmentByCandidateIdRoleIdAndAccomplishmentId } from "@/lib/db/resume/experience/accomplishment"
 import { getRoleByCandidateIdAndRoleId } from "@/lib/db/resume/experience/role"
